+++ deno.json
@@ -8,8 +8,7 @@
     "./sequencing": "./combinators/sequencing/mod.ts",
     "./primitives": "./primitives/mod.ts",
     "./common": "./common/mod.ts",
-    "./html": "./exports/html.ts",
-    "./csv": "./exports/csv.ts"
+    "./examples": "./examples/mod.ts"
   },
   "license": "MIT",
   "tasks": {
