--- exports/csv.ts
@@ -1,1 +1,0 @@
-export { csv } from "../examples/csv.ts";
