--- exports/html.ts
@@ -1,18 +1,0 @@
-export {
-  booleanAttributes,
-  commentNode,
-  fragments,
-  html,
-  Kind,
-  type MCommentNode,
-  type MElement,
-  type MFragment,
-  type MNode,
-  type MSpacesAndComments,
-  type MTextNode,
-  type SerializationOptions,
-  serializeFragments,
-  serializeNode,
-  shadowRoot,
-  textNode,
-} from "../examples/html.ts";
